requirements.txt
asttokens==3.0.0
black==25.9.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
greenlet==3.2.4
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
psycopg2==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
pywin32==311
pyzmq==27.1.0
six==1.17.0
SQLAlchemy==2.0.44
stack-data==0.6.3
tokenize_rt==6.2.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
